adding mt_deconv
Adding a wrapper to the mt_deconv function available in the fortran library.
Test doesn't pass, not sure of the result file I found in the examples, might need to run the routine directly and get the output in order to compare. (visually they are very similar!)
Doc should be OK. Tutorial has been updated to include the reproduction of Figure 6 from the original publication.