fix wrong confidence interval of mt_coherence()
In `multitaper.py`, the `mt_coherence()` function assigned an empty array
for the desired confidence interval using `mt.empty(nf, 2)` at line 600.
This gave an 1d complex array of length nf, because the argument `2` is passed
to the `complex=False` kwarg of the `mt.empty()` method.

The correct call should be `mt.empty((nf, 2))`